Issue #1: Allow formatter language embed settings
Embeddable JS app will have an feature that allows site to embed the
application with pre-defined language.

In Drupal, there are many ways to specify which language to pass on
to embedded app. We must have a way to configure this per formatter.

This change will create a setting for formatter that allows site
builders to specify the source of language to be passed.