Issue #1: Implement language attribute logic
We need to pass the language attribute to the JS application so that,
it can preselect an language that is already done by context web
site.

There is an formatter setting where site builder can decide where
this language is fetched.

This change will adapt this setting and run the logic for fetching
the language attribute according to formatter settings.