Issue #2: Embed foreign JS
Previously we were hosting the JS that was fetched through NPM
dependency as part of module. This decision was made based on known
fact that the script was still in experimental stage.

Now after taking the service into University of Helsinki's "student
services" official support, we can trust enough to embed it directly.